Add files via upload
I need help comparing the ISBN of a stored value to a newly entered ISBN so that the program can tell the user if that book is already stored...i.e. if ISBN 12345 Title Happy is stored and someone tries to store ISBN 12345 Title Angry It needs to return ISBN 12345 Title Happy already exist.